+++ src/main/java/goorm/humandelivery/domain/model/response/ErrorResponse.java
@@ -8,7 +8,6 @@ import lombok.Setter;
 import lombok.NoArgsConstructor;
 
 
-@Getter
 @AllArgsConstructor
 @Getter
 @Setter
